change logging so that messages can be pushed to WebSocket

+++ BrusaDMC5.h
@@ -160,7 +160,7 @@ public:
     void handleStateChange(Status::SystemState, Status::SystemState);
     void tearDown();
     DeviceId getId();
-    int16_t getMechanicalPower();
+    int32_t getMechanicalPower();
 
     void loadConfiguration();
     void saveConfiguration();
